+++ com.unity.robotics.urdf-importer/CHANGELOG.md
@@ -21,7 +21,7 @@ Add capsule shape support
 - Removed option on UrdfRobot inspector window to set all MeshColliders to non-Convex meshes.
 
 ### Fixed
-
+- Added check for NaN values during inertia matrix's conversion to inertia tensor rotation.
 
 ## [0.5.2-preview] - 2022-02-01